requirements.txt
# Core LiteLLM dependencies
litellm[proxy]==1.47.0
uvicorn[standard]==0.24.0
fastapi==0.104.1

# Database support
sqlalchemy==2.0.23
alembic==1.12.1

# Redis for caching (optional)
redis==5.0.1
hiredis==2.2.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration management
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client improvements
httpx==0.25.2
aiohttp==3.9.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional provider-specific dependencies
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
cohere==4.37
boto3==1.34.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
